+++ service/internal/workflow/workflow.go
@@ -20,7 +20,7 @@ const mainContainer = "main"
 
 // Workflow interface is used for interacting with workflow services.
 type Workflow interface {
-	List(ctx context.Context) ([]string, error)
+	ListStatus(ctx context.Context) ([]Status, error)
 	Logs(ctx context.Context, workflowName string) (*Logs, error)
 	LogStream(ctx context.Context, workflowName string, data http.ResponseWriter) error
 	Status(ctx context.Context, workflowName string) (*Status, error)
@@ -46,23 +46,32 @@ type Logs struct {
 	Logs []string `json:"logs"`
 }
 
-// List returns a list of workflows.
-func (a ArgoWorkflow) List(ctx context.Context) ([]string, error) {
-	workflowIDs := []string{}
-
+// List returns a list of workflow statuses.
+func (a ArgoWorkflow) ListStatus(ctx context.Context) ([]Status, error) {
 	workflowListResult, err := a.svc.ListWorkflows(ctx, &argoWorkflowAPIClient.WorkflowListRequest{
 		Namespace: a.namespace,
 	})
-
 	if err != nil {
-		return workflowIDs, err
+		return []Status{}, err
 	}
 
-	for _, item := range workflowListResult.Items {
-		workflowIDs = append(workflowIDs, item.ObjectMeta.Name)
+	workflows := make([]Status, len(workflowListResult.Items))
+
+	for k, wf := range workflowListResult.Items {
+		wfStatus := Status{
+			Name:    wf.ObjectMeta.Name,
+			Status:  strings.ToLower(string(wf.Status.Phase)),
+			Created: fmt.Sprint(wf.ObjectMeta.CreationTimestamp.Unix()),
+		}
+
+		if wf.Status.Phase != argoWorkflowAPISpec.WorkflowRunning {
+			wfStatus.Finished = fmt.Sprint(wf.Status.FinishedAt.Unix())
+		}
+
+		workflows[k] = wfStatus
 	}
 
-	return workflowIDs, nil
+	return workflows, nil
 }
 
 // Status represents a workflow status.
@@ -70,7 +79,7 @@ type Status struct {
 	Name     string `json:"name"`
 	Status   string `json:"status"`
 	Created  string `json:"created"`
-	Finished string `json:"finished"`
+	Finished string `json:"finished,omitempty"`
 }
 
 // Status returns a workflow status.
@@ -79,7 +88,6 @@ func (a ArgoWorkflow) Status(ctx context.Context, workflowName string) (*Status,
 		Name:      workflowName,
 		Namespace: a.namespace,
 	})
-
 	if err != nil {
 		return nil, err
 	}
@@ -103,7 +111,6 @@ func (a ArgoWorkflow) Logs(ctx context.Context, workflowName string) (*Logs, err
 			Container: mainContainer,
 		},
 	})
-
 	if err != nil {
 		return nil, err
 	}
@@ -135,7 +142,6 @@ func (a ArgoWorkflow) LogStream(ctx context.Context, workflowName string, w http
 			Follow:    true,
 		},
 	})
-
 	if err != nil {
 		return err
 	}
@@ -197,7 +203,6 @@ func (a ArgoWorkflow) Submit(ctx context.Context, from string, parameters map[st
 			Labels:       labels.FormatLabels(workflowLabels),
 		},
 	})
-
 	if err != nil {
 		return "", fmt.Errorf("failed to submit workflow: %w", err)
 	}